refactor(tui): extract input style and placeholder component
